requirements.txt
streamlit
pandas
streamlit_extras
pathlib
altair
plotly.express